// src/models/Customer.tsx
interface Customer {
    _id: string
    FirstName: string
    LastName: string
    Gender: string
    DOB: any
    PhoneNumber: string
    Email: string
    GoogleId: string
    Gmail: string
    EmergencyTel?: string 
    Avatar?: any
    CongenitalDisorders?: string
    Role: string
    CreatedAt: any
    UpdatedAt: any
}

export default Customer